Change tabs to the tabs components of the dress code. Rename dress-code file.
