+++ web-ui/.babelrc
@@ -1,3 +1,4 @@
 {
-  "presets" : ["es2015", "react"]
+  "presets" : ["es2015", "react"],
+  "plugins": ["transform-object-rest-spread"]
 }
