+++ web-ui/package.json
@@ -32,7 +32,8 @@
   "homepage": "https://github.com/zalando-incubator/zally/web-ui#readme",
   "dependencies": {
     "aurelia-fetch-client": "1.1.2",
-    "dress-code": "2.0.0",
+    "classnames": "2.2.5",
+    "dress-code": "2.1.1",
     "express": "4.15.2",
     "history": "3.0.0",
     "lodash": "4.17.4",
@@ -41,12 +42,14 @@
     "react-ace": "5.2.0",
     "react-dom": "15.4.2",
     "react-router": "3.0.4",
+    "react-router-bootstrap": "0.23.3",
     "request": "2.81.0",
     "whatwg-fetch": "2.0.3"
   },
   "devDependencies": {
     "babel-core": "6.24.0",
     "babel-loader": "6.4.0",
+    "babel-plugin-transform-object-rest-spread": "6.26.0",
     "babel-preset-es2015": "6.24.0",
     "babel-preset-react": "6.23.0",
     "codecov": "2.1.0",
