+++ web-ui/src/client/app/containers/rules.jsx
@@ -80,7 +80,7 @@ export class Rules extends Component {
         <div className="dc-row">
           <div className="dc-column dc-column--small-12 dc-column--large-7">
             <div className="dc-column__contents">
-              <RulesTab error={this.state.error} rules={this.state.rules}></RulesTab>
+              <RulesTab error={this.state.error} rules={this.state.rules}/>
             </div>
           </div>
         </div>
