+++ web-ui/src/server/views/index.pug
@@ -5,7 +5,7 @@ html
     base(href=mountpath)
     meta(name="viewport" content="width=device-width, initial-scale=1")
     link(href="https://fonts.googleapis.com/css?family=Ubuntu:300,500,300italic|Ubuntu+Mono" rel="stylesheet" type="text/css")
-    link(href="https://unpkg.com/dress-code@2.0.0/dist/css/dress-code.min.css" rel="stylesheet" type="text/css")
+    link(href="https://unpkg.com/dress-code@2.1.1/dist/css/dress-code.min.css" rel="stylesheet" type="text/css")
     link(rel="apple-touch-icon-precomposed", sizes="57x57", href=mountpath + "assets/icons/apple-touch-icon-57x57.png")
     link(rel="apple-touch-icon-precomposed", sizes="114x114", href=mountpath + "assets/icons/apple-touch-icon-114x114.png")
     link(rel="apple-touch-icon-precomposed", sizes="72x72", href=mountpath + "assets/icons/apple-touch-icon-72x72.png")
